solver:forced-moves

+++ branches/pure-lw1/src/new_solver.h
@@ -107,7 +107,7 @@ int NewSolver<T>::solve(const T &initial_hidden_state,
     Instance::Plan raw_plan, plan;
     final_plan.clear();
 
-    size_t planner_calls = 0;
+    int planner_calls = 0;
     while( !state.goal(kp_instance_) ) {
 
         plan.clear();
